fix: moved testUtils.js
