+++ frontend/src/tests/unit/components/AuthView/AuthForm.test.js
@@ -4,7 +4,7 @@ import { mount } from "@vue/test-utils";
 import AuthForm from "@/components/AuthView/AuthForm.vue";
 import { AuthFormText } from "@/constants/AuthFormText";
 import AuthService from "@/services/AuthService";
-import { createCorrectUserSignUp } from "../../utils/testUtils";
+import { createCorrectUserSignUp } from "../../../utils/testUtils";
 
 
 describe("AuthForm", () => {
